Dev dim n valid num refine (#1336)
* dimN_valid_num=>dimN_valid_num_ptr

* dimN_valid_num default value

* int32_t=>int64_t

* refine raw encoder

+++ oneflow/core/record/ofrecord_raw_decoder.cpp
@@ -23,7 +23,7 @@ bool OFRecordDecoderImpl<EncodeCase::kRaw, T>::HasDim1ValidNumField(
 template<typename T>
 void OFRecordDecoderImpl<EncodeCase::kRaw, T>::SetDim1ValidNum(const Feature& feature,
                                                                Blob* out_blob,
-                                                               int32_t dim0_idx) const {
+                                                               int64_t dim0_idx) const {
   CHECK_GE(out_blob->static_shape().NumAxes(), 2);
   int64_t elem_num = 0;
   if (feature.has_bytes_list()) {
